+++ QiniuSDK/Storage/QNUploadInfoV2.m
@@ -193,7 +193,7 @@
         }
         
         // 从资源中读取新的 block 进行上传
-        long dataOffset = 0;
+        long long dataOffset = 0;
         if (self.dataList.count > 0) {
             QNUploadData *lastData = self.dataList[self.dataList.count - 1];
             dataOffset = lastData.offset + lastData.size;
@@ -273,7 +273,7 @@
     
     // 未加载过 block 数据
     // 根据 data 信息加载 dataBytes
-    NSData *dataBytes = [self readData:data.size dataOffset:data.offset error:error];
+    NSData *dataBytes = [self readData:(NSInteger)data.size dataOffset:data.offset error:error];
     if (*error != nil) {
         return nil;
     }
