+++ QiniuSDK/Storage/QNUploadSourceStream.m
@@ -12,7 +12,7 @@
 @interface QNUploadSourceStream()
 
 @property(nonatomic, assign)long long size;
-@property(nonatomic, assign)NSInteger readOffset;
+@property(nonatomic, assign)long long readOffset;
 @property(nonatomic,   copy)NSString *sourceId;
 @property(nonatomic,   copy)NSString *fileName;
 @property(nonatomic, strong)NSInputStream *stream;
@@ -57,7 +57,7 @@
     }
 }
 
-- (NSData *)readData:(NSInteger)dataSize dataOffset:(long)dataOffset error:(NSError **)error {
+- (NSData *)readData:(NSInteger)dataSize dataOffset:(long long)dataOffset error:(NSError **)error {
     if (self.stream == nil) {
         *error = [NSError errorWithDomain:NSCocoaErrorDomain code:kQNFileError userInfo:@{NSLocalizedDescriptionKey : @"inputStream is empty"}];
         return nil;
@@ -139,7 +139,7 @@
     }
 }
 
-- (void)streamSkipSize:(NSInteger)size error:(NSError **)error {
+- (void)streamSkipSize:(long long)size error:(NSError **)error {
     BOOL isEOF = false;
     NSInteger sliceSize = 1024;
     NSInteger readSize = 0;
