+++ QiniuSDK/Utils/QNALAssetFile.m
@@ -43,7 +43,7 @@
     return self;
 }
 
-- (NSData *)read:(long)offset
+- (NSData *)read:(long long)offset
             size:(long)size
            error:(NSError **)error {
